requirements.txt
Flask==2.3.2
Flask_Bcrypt==1.0.1
Flask_Login==0.6.2
Flask_Mail==0.9.1
flask_sqlalchemy==3.1.1
Flask_WTF==1.1.1
Pillow==9.5.0
Pillow==10.0.1
WTForms==3.0.1
